+++ app/src/main/java/com/eronalves/tictactoe/data/Move.kt
@@ -8,11 +8,11 @@ import androidx.room.Relation
 
 @Entity
 data class Move(
-    @PrimaryKey(autoGenerate = true) val id: Int?,
+    @PrimaryKey(autoGenerate = true) val id: Long?,
     @ColumnInfo(name = "t_row") val row: Int?,
     @ColumnInfo(name = "t_column") val column: Int?,
-    @ColumnInfo(name = "player_id") val playerId: Int?,
-    @ColumnInfo(name = "round_id") val roundId: Int?
+    @ColumnInfo(name = "player_id") val playerId: Long?,
+    @ColumnInfo(name = "round_id") val roundId: Long?
 )
 
 data class MovePlayerAndRound(
